feat(P-2): renamed definitions directory to types
