+++ src/features/Sentence/Sentence.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react';
 import React, { useState } from 'react'
-import { TElementProps } from '../../definitions';
+import { TElementProps } from '../../types';
 import { CWrap, Last, WWrap } from './Sentence.styles';
 
 export type IProps = TElementProps;
