+++ src/theme/components/Count/cwrap.ts
@@ -1,4 +1,4 @@
-import { TCSSObjProps } from "../../../definitions";
+import { TCSSObjProps } from "../../../types";
 import { show } from "./keyframes";
 
 export const blocks: TCSSObjProps = {
