+++ src/theme/components/Sentence/last.ts
@@ -1,4 +1,4 @@
-import { TCSSObjProps } from "../../../definitions";
+import { TCSSObjProps } from "../../../types";
 import { lastWord } from "./keyframes";
 
 export const blocks: TCSSObjProps = {
