+++ src/types/TCSSObjProps.ts
