+++ src/types/TMouseCoords.ts
